# store: read postings directly into diffvarint format