… .bingo/promdoc.sum