Instead of allocating bytes for raw postings, let's read them directly into diff varint format to save memory. ~This should already cut memory usage by 3x.~ Savings are smaller. See comment.