@yeya24 2023-06-15
QQ @GiedriusS, this doesn't solve https://github.com/thanos-io/thanos/issues/6434, right?